Update molecule for newer OS releases
Currently the molecule CI fails because they renamed a package, and it
doesn't work as expected with the cgroups changes in Ubuntu 22.04.

We should lock GH Actions at 22.04 for a base runtime, and bump all the
test platforms to their latest stable releases as well. Also python
since we're at it.